+++ lib/widgets/usuarios_card.dart
@@ -145,7 +145,7 @@ class _TerminoState extends State<_Termino> {
 }
 
 class _PriceTag extends StatefulWidget {
-  final int price;
+  final double price;
 
   const _PriceTag(this.price);
 
@@ -214,6 +214,7 @@ class _ProductDetailsState extends State<_ProductDetails> {
               style: const TextStyle(fontSize: 11, color: Colors.white),
             ),
             Text(
+              //'vs granja de: ${widget.creador[0].toUpperCase()}${widget.creador.substring(1).split('@')[0]}',
               'vs granja de: ${widget.creador[0].toUpperCase()}${widget.creador.substring(1).split('@')[0]}',
               //${rrvalue![0].toUpperCase()}${rrvalue.substring(1).split('@')[0]}
               style: const TextStyle(fontSize: 11, color: Colors.white),
